+++ packages/core/List/index.js
@@ -21,6 +21,7 @@ const logger = require('@voussoir/logger');
 
 const { Text, Checkbox, Float, Relationship } = require('@voussoir/fields');
 
+const gql = require('graphql-tag');
 const graphqlLogger = logger('graphql');
 const keystoneLogger = logger('keystone');
 
@@ -502,7 +503,13 @@ module.exports = class List {
   get gqlAuxMutationResolvers() {
     // TODO: Obey the same ACL rules based on parent type
     return objMerge([
-      this.config.gqlAuxMutationResolvers,
+      ...(this.config.mutations || []).map(({ schema, resolver }) => {
+        const mutationName = gql(`type t { ${schema} }`).definitions[0].fields[0].name.value;
+        return {
+          [mutationName]: (obj, args, context, info) =>
+            resolver(obj, args, context, info, mapKeys(this.hooksActions, hook => hook(context))),
+        };
+      }),
       ...this.fields.map(field => field.gqlAuxMutationResolvers),
     ]);
   }
@@ -511,8 +518,8 @@ module.exports = class List {
     const mutations = flatten(
       this.fields.map(field => field.getGqlAuxMutations({ skipAccessControl }))
     );
-    if (this.config.getGqlAuxMutations) {
-      mutations.push(this.config.getGqlAuxMutations({ skipAccessControl }));
+    if (this.config.mutations) {
+      mutations.push(...this.config.mutations.map(({ schema }) => schema));
     }
 
     // NOTE: We only check for truthy as it could be `true`, or a function (the